Add small note about CPU/mem usage
My main concern to run this in my production k8s cluster was it's
CPU/mem usage. So, for example, if one node dies I was afraid this pod
might get scheduled and not others, more important, pods.

But, looking at it, it's really lightweight and can fit in any node just
fine without causing any issues (nor having more nodes). So, it may be
useful for other to see there is really no risk.